Consolidate util modules in their own dir
