+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './codeprinter';
 export * from './producers/producer';
 export * from './receivers/renderer';
 export * from './types';
+export * from './util/codeprinter';
