+++ src/producers/barlinehandler.test.ts
@@ -1,7 +1,7 @@
-import * as msg from '../msg';
 import { NoopReceiver } from '../receivers/noopreceiver';
 import { BarlineMessage } from '../types';
-import * as xml from '../xml';
+import * as msg from '../util/msg';
+import * as xml from '../util/xml';
 import { BarlineHandler } from './barlinehandler';
 
 describe('BarlineHandler', () => {
