+++ src/producers/measurehandler.ts
@@ -1,7 +1,7 @@
-import * as msg from '../msg';
-import { NamedNode } from '../namednode';
 import { MultiReceiver } from '../receivers/multireceiver';
 import { VexmlConfig, VexmlMessageReceiver } from '../types';
+import * as msg from '../util/msg';
+import { NamedNode } from '../util/namednode';
 import { NodeHandler, NodeHandlerCtx } from './nodehandler';
 
 export class MeasureHandlerError extends Error {}
