+++ src/producers/parthandler.test.ts
@@ -1,7 +1,7 @@
 import { DEFAULT_CONFIG } from '../di';
 import { NoopReceiver } from '../receivers/noopreceiver';
 import { PartEndMessage, PartStartMessage } from '../types';
-import * as xml from '../xml';
+import * as xml from '../util/xml';
 import { NoopHandler } from './noophandler';
 import { PartHandler } from './parthandler';
