+++ src/util/partiterator.ts
